Avoid compiler warnings due to differences in signs in comparisons

+++ lib/libebml/src/EbmlElement.cpp
@@ -49,7 +49,7 @@ START_LIBEBML_NAMESPACE
 */
 int CodedSizeLength(uint64 Length, unsigned int SizeLength, bool bSizeFinite)
 {
-	int CodedSize;
+	unsigned int CodedSize;
 	if (bSizeFinite) {
 		// prepare the head of the size (000...01xxxxxx)
 		// optimal size
@@ -587,7 +587,7 @@ filepos_t EbmlElement::Render(IOCallback & output, bool bWithDefault, bool bKeep
 		filepos_t result = RenderHead(output, bForceRender, bWithDefault, bKeepPosition);
 		uint64 WrittenSize = RenderData(output, bForceRender, bWithDefault);
 #if defined(_DEBUG) || defined(DEBUG)
-	if (SupposedSize != (0-1)) assert(WrittenSize == SupposedSize);
+	if (static_cast<int64_t>(SupposedSize) != (0-1)) assert(WrittenSize == SupposedSize);
 #endif // DEBUG
 		result += WrittenSize;
 		return result;
